requirements.txt
torch
torchvision
torchaudio
opencv-python
numpy
pillow
tqdm
matplotlib
pywavelets
scikit-image
torchmetrics
streamlit
